🐛 Fix application quitting on return in submenu if entered during loading transition
GameManager.LoadManagedSceneAsync would only return after the full transition. As a consequence LoadMenuAsync would only subscribe to the menu manager events after the transition was over, which could be after getting in a menu. As a consequence, HandleEnterSubmenu would not be invoked on time.

The solution was to split the loading and transitioning parts, so scene initialization and subscription can happen as soon as possible.