🐛 Fix application quitting when pressing back button in a match
GameManager.HandleReturn called LoadMenuAsync right away, and its first line of code was setting Input.backButtonLeavesApp to true. This call was triggered by the back button, and setting Input.backButtonLeavesApp to true inside it seems to cause the application quit. The back button press was most likely handled after user code. The fix was to wait until the menu is loaded to set Input.backButtonLeavesApp to true.